feat: migrate to Vite + Supabase setup Switched project build system to Vite Integrated Supabase database to manage quotes
