+++ index.html
@@ -25,16 +25,38 @@
     />
 
     <!-- Favicon -->
-    <link rel="apple-touch-icon" sizes="180x180" href="./favicon/apple-touch-icon.png" />
-    <link rel="icon" type="image/png" sizes="32x32" href="./favicon/favicon-32x32.png" />
-    <link rel="icon" type="image/png" sizes="16x16" href="./favicon/favicon-16x16.png" />
-    <link rel="icon" href="./favicon/favicon.ico">
-    <link rel="icon" href="./favicon/android-chrome-192x192.png" sizes="192x192" />
-    <link rel="icon" href="./favicon/android-chrome-512x512.png" sizes="512x512" />
-    <link rel="manifest" href="/site.webmanifest" />
+    <link
+      rel="apple-touch-icon"
+      sizes="180x180"
+      href="/public/favicon/apple-touch-icon.png"
+    />
+    <link
+      rel="icon"
+      type="image/png"
+      sizes="32x32"
+      href="public/favicon/favicon-32x32.png"
+    />
+    <link
+      rel="icon"
+      type="image/png"
+      sizes="16x16"
+      href="public/favicon/favicon-16x16.png"
+    />
+    <link rel="icon" href="public/favicon/favicon.ico" />
+    <link
+      rel="icon"
+      href="public/favicon/android-chrome-192x192.png"
+      sizes="192x192"
+    />
+    <link
+      rel="icon"
+      href="public/favicon/android-chrome-512x512.png"
+      sizes="512x512"
+    />
+    <link rel="manifest" href="public/site.webmanifest" />
 
     <!-- css file -->
-    <link rel="stylesheet" href="styles.css" />
+    <link rel="stylesheet" href="/src/styles.css" />
   </head>
 
   <body>
@@ -58,19 +80,19 @@
       <section class="quote-section">
         <div class="quote-container">
           <p class="quote-text" id="quoteText">
-            Click "Inspire Me" to discover your next favorite quote!
+            The best way to cheer yourself up is to try to cheer somebody else up.
           </p>
-          <p class="quote-author" id="quoteAuthor">InspireMe</p>
+          <p class="quote-author" id="quoteAuthor">Mark Twain</p>
         </div>
         <button
           class="btn"
           id="newQuote"
           aria-label="Get new inspirational quote"
         >
-          ✨ New Quote
+          New Quote
         </button>
       </section>
-      
+
       <footer>
         <p>
           <span class="sr-only-year" id="sr-year"></span>
@@ -105,7 +127,6 @@
                 d="M20.447 20.452h-3.554v-5.569c0-1.328-.027-3.037-1.852-3.037-1.853 0-2.136 1.445-2.136 2.939v5.667H9.351V9h3.414v1.561h.046c.477-.9 1.637-1.85 3.37-1.85 3.601 0 4.267 2.37 4.267 5.455v6.286zM5.337 7.433c-1.144 0-2.063-.926-2.063-2.065 0-1.138.92-2.063 2.063-2.063 1.14 0 2.064.925 2.064 2.063 0 1.139-.925 2.065-2.064 2.065zm1.782 13.019H3.555V9h3.564v11.452zM22.225 0H1.771C.792 0 0 .774 0 1.729v20.542C0 23.227.792 24 1.771 24h20.451C23.2 24 24 23.227 24 22.271V1.729C24 .774 23.2 0 22.222 0h.003z"
               />
             </svg>
-            
           </a>
           <a
             href="https://github.com/chiomavera/"
@@ -121,8 +142,8 @@
               <path
                 d="M12 0c-6.626 0-12 5.373-12 12 0 5.302 3.438 9.8 8.207 11.387.599.111.793-.261.793-.577v-2.234c-3.338.726-4.033-1.416-4.033-1.416-.546-1.387-1.333-1.756-1.333-1.756-1.089-.745.083-.729.083-.729 1.205.084 1.839 1.237 1.839 1.237 1.07 1.834 2.807 1.304 3.492.997.107-.775.418-1.305.762-1.604-2.665-.305-5.467-1.334-5.467-5.931 0-1.311.469-2.381 1.236-3.221-.124-.303-.535-1.524.117-3.176 0 0 1.008-.322 3.301 1.23.957-.266 1.983-.399 3.003-.404 1.02.005 2.047.138 3.006.404 2.291-1.552 3.297-1.23 3.297-1.23.653 1.653.242 2.874.118 3.176.77.84 1.235 1.911 1.235 3.221 0 4.609-2.807 5.624-5.479 5.921.43.372.823 1.102.823 2.222v3.293c0 .319.192.694.801.576 4.765-1.589 8.199-6.086 8.199-11.386 0-6.627-5.373-12-12-12z"
               />
-              </svg>
-            </a>
+            </svg>
+          </a>
           <a
             href="https://twitter.com/nkanmuo_vera"
             target="_blank"
@@ -144,6 +165,7 @@
     </div>
 
     <!-- JavaScript file -->
-    <script type="module" defer src="main.js"></script>
+    <div id="app"></div>
+    <script type="module" src="/src/main.js"></script>
   </body>
 </html>
