+++ package.json
@@ -1,0 +1,18 @@
+{
+  "name": "inspireme",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "vite": "^7.1.7"
+  },
+  "dependencies": {
+    "@supabase/supabase-js": "^2.76.0",
+    "tslib": "^2.8.1"
+  }
+}
