--- quotes.js
@@ -1,197 +1,0 @@
-const quotes = [
-    // Inspiration & Motivation
-    {
-      text: "The journey of a thousand miles begins with one step.",
-      author: "Lao Tzu",
-      category: "Inspiration & Motivation"
-    },
-    {
-      text: "Believe you can and you're halfway there.",
-      author: "Theodore Roosevelt",
-      category: "Inspiration & Motivation"
-    },
-    {
-      text: "Your time is limited, so don't waste it living someone else's life.",
-      author: "Steve Jobs",
-      category: "Inspiration & Motivation"
-    },
-    {
-      text: "It does not matter how slowly you go as long as you do not stop.",
-      author: "Confucius",
-      category: "Inspiration & Motivation"
-    },
-    {
-      text: "Success is not final, failure is not fatal: It is the courage to continue that counts.",
-      author: "Winston Churchill",
-      category: "Inspiration & Motivation"
-    },
-    {
-      text: "Do what you feel in your heart to be right – for you'll be criticized anyway.",
-      author: "Eleanor Roosevelt",
-      category: "Inspiration & Motivation"
-    },
-    { 
-      text: "Don't count the days, make the days count.",
-      author: "Muhammad Ali",
-      category: "Inspiration & Motivation"
-    },
-    {
-      text: "Hardships often prepare ordinary people for an extraordinary destiny.",
-      author: "C.S. Lewis",
-      category: "Inspiration & Motivation"
-    },
-    {
-      text: "A river cuts through rock, not because of its power, but because of its persistence.",
-      author: "James N. Watkins",
-      category: "Inspiration & Motivation"
-    },
-  
-    // Wisdom & Philosophy
-    {
-      text: "An unexamined life is not worth living.",
-      author: "Socrates",
-      category: "Wisdom & Philosophy"
-    },
-    {
-      text: "Happiness is the highest good.",
-      author: "Aristotle",
-      category: "Wisdom & Philosophy"
-    },
-    {
-      text: "The only true wisdom is in knowing you know nothing.",
-      author: "Socrates",
-      category: "Wisdom & Philosophy"
-    },
-    {
-      text: "Real integrity is doing the right thing, even when no one is watching.",
-      author: "C.S. Lewis",
-      category: "Wisdom & Philosophy"
-    },
-    {
-      text: "Tell me and I forget. Teach me and I remember. Involve me and I learn.",
-      author: "Benjamin Franklin",
-      category: "Wisdom & Philosophy"
-    },
-    {
-      text: "The beautiful thing about learning is that no one can take it away from you.",
-      author: "B.B. King",
-      category: "Wisdom & Philosophy"
-    },
-    {
-      text: "An investment in knowledge pays the best interest.",
-      author: "Benjamin Franklin",
-      category: "Wisdom & Philosophy"
-    },
-    {
-      text: "The greatest wealth is to live content with little.",
-      author: "Plato",
-      category: "Wisdom & Philosophy"
-    },
-    {
-      text: "He who opens a school door, closes a prison.",
-      author: "Victor Hugo",
-      category: "Wisdom & Philosophy"
-    },
-  
-    // Happiness & Gratitude
-    {
-      text: "If you want to live a happy life, tie it to a goal, not to people or things.",
-      author: "Albert Einstein",
-      category: "Happiness & Gratitude"
-    },
-    {
-      text: "Happiness depends upon ourselves.",
-      author: "Aristotle",
-      category: "Happiness & Gratitude"
-    },
-    {
-      text: "The purpose of our lives is to be happy.",
-      author: "Dalai Lama",
-      category: "Happiness & Gratitude"
-    },
-    {
-      text: "If you look at what you have in life, you'll always have more.",
-      author: "Oprah Winfrey",
-      category: "Happiness & Gratitude"
-    },
-    {
-      text: "Gratitude turns what we have into enough.",
-      author: "Melody Beattie",
-      category: "Happiness & Gratitude"
-    },
-    {
-      text: "The root of joy is gratefulness.",
-      author: "David Steindl-Rast",
-      category: "Happiness & Gratitude"
-    },
-    {
-      text: "Happiness is not something ready made. It comes from your own actions.",
-      author: "Dalai Lama",
-      category: "Happiness & Gratitude"
-    },
-    {
-      text: "Enjoy the little things, for one day you may look back and realize they were the big things.",
-      author: "Robert Brault",
-      category: "Happiness & Gratitude"
-    },
-    {
-      text: "The best way to cheer yourself up is to try to cheer somebody else up.",
-      author: "Mark Twain",
-      category: "Happiness & Gratitude"
-    },
-  
-    // Humor & Service
-    {
-      text: "Life would be a great deal easier if dead things had the decency to remain dead.",
-      author: "Doug MacLeod",
-      category: "Humor & Service"
-    },
-    {
-      text: "I am so clever that sometimes I don't understand a single word of what I am saying.",
-      author: "Oscar Wilde",
-      category: "Humor & Service"
-    },
-    {
-      text: "The only mystery in life is why the kamikaze pilots wore helmets.",
-      author: "Al McGuire",
-      category: "Humor & Service"
-    },
-    {
-      text: "No one has ever become poor by giving.",
-      author: "Anne Frank",
-      category: "Humor & Service"
-    },
-    {
-      text: "We make a living by what we get, but we make a life by what we give.",
-      author: "Winston Churchill",
-      category: "Humor & Service"
-    },
-    {
-      text: "The best way to find yourself is to lose yourself in the service of others.",
-      author: "Mahatma Gandhi",
-      category: "Humor & Service"
-    },
-    {
-      text: "Injustice anywhere is a threat to justice everywhere.",
-      author: "Martin Luther King, Jr.",
-      category: "Humor & Service"
-    },
-    {
-      text: "The time is always right to do what is right.",
-      author: "Martin Luther King, Jr.",
-      category: "Humor & Service"
-    },
-    {
-      text: "At his best, man is the noblest of all animals; separated from law and justice he is the worst.",
-      author: "Aristotle",
-      category: "Humor & Service"
-    },
-    {
-      text: "Kindness is a language which the deaf can hear and the blind can see.",
-      author: "Mark Twain",
-      category: "Humor & Service"
-    }
-  ];
-  
-  export default quotes;
-
